% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyp_box.R
\name{polyp_box}
\alias{polyp_box}
\title{poly_box}
\usage{
polyp_box(data, group, value)
}
\arguments{
\item{data}{Input data}

\item{group}{Group variable in the data}

\item{value}{Numeric variable in the data}
}
\description{
poly_box
}
\examples{
polyp_box(poly, "trt", "number")
}
