% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyp_estdp.R
\name{polyp_estdp}
\alias{polyp_estdp}
\title{Poisson model summary with estimated dispersion parameter}
\usage{
polyp_estdp(data, group, Y, value)
}
\arguments{
\item{data}{Input data}

\item{group}{Group variable in the data}

\item{Y}{Numeric variable in the data}

\item{value}{Numeric variable in the data}
}
\description{
Poisson model summary with estimated dispersion parameter
}
\examples{
polyp_estdp(data=poly, group="trt", Y="number", value="age")

}
