% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyp_hist.R
\name{polyp_hist}
\alias{polyp_hist}
\title{poly_hist}
\usage{
polyp_hist(data, group, value)
}
\arguments{
\item{data}{Input data}

\item{group}{Group variable in the data}

\item{value}{Numeric variable in the data}
}
\description{
poly_hist
}
\examples{
polyp_hist(poly, "trt", "number")
}
