% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyp_model.R
\name{polyp_model}
\alias{polyp_model}
\title{polyp_model
Description: select proper distribution based on AIC, and present a summary of distribution}
\usage{
polyp_model(data, group, Y, value)
}
\arguments{
\item{data}{Input data}

\item{group}{Group}

\item{Y}{Numeric variable}

\item{value}{Numeric variable}
}
\description{
polyp_model
Description: select proper distribution based on AIC, and present a summary of distribution
}
\examples{
polyp_model(data = poly, group = "trt", Y = "number", value = "age")

}
