% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyp_stat.R
\name{polyp_stat}
\alias{polyp_stat}
\title{polyp_stat}
\usage{
polyp_stat(data, value)
}
\arguments{
\item{data}{Input data}

\item{value}{Numeric variable in the data}

\item{group}{Group variable in the data}
}
\description{
polyp_stat
}
\examples{
polyp_stat(poly, "number")
}
