% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_plot.R
\name{pre_plot}
\alias{pre_plot}
\title{Predicted values plot about final model}
\usage{
pre_plot(data, group, Y, value)
}
\arguments{
\item{data}{Input data}

\item{group}{Group}

\item{Y}{Numeric variable}

\item{value}{Numeric variable}
}
\description{
Predicted values plot about final model
}
\examples{
pre_plot(poly,"trt","number","age")

}
